refactor: move data hydration logic to computed stores from the components, move getChildHcRows to helpers
