+++ src/components/charsheet/equipment/cyber/cyberViewTypes.ts
@@ -1,40 +1,3 @@
-import type { CyberCategory } from "@scripts/cyber/catalog";
-import type { HydratedCyberItem } from "@stores/cyber";
-
-export interface CyberItem {
-  id: string;
-  name: string;
-  category: CyberCategory;
-  description: string;
-  hc: number | string;
-  owned: boolean;
-  installed: boolean;
-  availability?: string;
-  cost?: number;
-  isBase?: boolean;
-  role?: "container" | "option" | "standalone";
-  parentId?: string;
-  slotUsage?: { used: number; max: number | null };
-  installedOptions?: string[];
-}
-
-export function hydratedToCyberItem(h: HydratedCyberItem): CyberItem {
-  return {
-    id: h.instanceId,
-    name: h.template.name,
-    category: h.template.category,
-    description: h.template.description,
-    hc: h.hc,
-    owned: true,
-    installed: h.installed,
-    availability: h.template.availability,
-    cost: h.template.cost,
-    isBase: h.template.role === "container",
-    role: h.template.role,
-    parentId: h.parentId,
-  };
-}
-
 export interface CyberlimbCell {
   slot: "la" | "ra" | "ll" | "rl";
   label: string;
