+++ src/components/charsheet/equipment/cyber/hooks/useInstallHelpers.ts
@@ -1,8 +1,7 @@
 import { useStore } from "@nanostores/preact";
 import { useMemo } from "preact/hooks";
 
-import { CYBER_CATALOG } from "@scripts/cyber/catalog";
-import { $hydratedCyber, $ownedCyber, getContainersForOption } from "@stores/cyber";
+import { $hydratedCyber, getContainersForOption } from "@stores/cyber";
 import { $homerules } from "@stores/homerules";
 
 // --- Resolve effective HC notation (respects tsmFreeHc houserule) ---
@@ -52,19 +51,4 @@ export function useContainerChoices(
       };
     });
   }, [templateId, hydrated]);
-}
-
-// --- Get children of a container for multi-row HC ---
-
-export function getChildHcRows(
-  containerId: string,
-): { key: string; name: string; notation: string }[] {
-  const items = $ownedCyber.get();
-  return items
-    .filter((i) => i.parentId === containerId)
-    .flatMap((i) => {
-      const t = CYBER_CATALOG[i.templateId];
-      if (!t) return [];
-      return [{ key: i.instanceId, name: t.name, notation: t.hc }];
-    });
 }
